+++ sdk/js/src/index.ts
@@ -5,3 +5,4 @@ export * from "./rpc";
 export * from "./utils";
 export * from "./bridge";
 export * from "./token_bridge";
+// TODO(csongor): should we export nft_bridge here? if so, the names will clash