+++ sdk/js/src/nft_bridge/transfer.ts
@@ -1,5 +1,6 @@
 import { Token, TOKEN_PROGRAM_ID } from "@solana/spl-token";
 import { Connection, Keypair, PublicKey, Transaction } from "@solana/web3.js";
+import { MsgExecuteContract } from "@terra-money/terra.js";
 import { ethers } from "ethers";
 import {
   NFTBridge__factory,
@@ -16,7 +17,7 @@ export async function transferFromEth(
   tokenID: ethers.BigNumberish,
   recipientChain: ChainId,
   recipientAddress: Uint8Array
-) {
+): Promise<ethers.ContractReceipt> {
   //TODO: should we check if token attestation exists on the target chain
   const token = NFTImplementation__factory.connect(tokenAddress, signer);
   await (await token.approve(tokenBridgeAddress, tokenID)).wait();
@@ -44,7 +45,7 @@ export async function transferFromSolana(
   originAddress?: Uint8Array,
   originChain?: ChainId,
   originTokenId?: Uint8Array
-) {
+): Promise<Transaction> {
   const nonce = createNonce().readUInt32LE(0);
   const transferIx = await getBridgeFeeIx(
     connection,
@@ -106,4 +107,42 @@ export async function transferFromSolana(
   transaction.feePayer = new PublicKey(payerAddress);
   transaction.partialSign(messageKey);
   return transaction;
+}
+
+export async function transferFromTerra(
+  walletAddress: string,
+  tokenBridgeAddress: string,
+  tokenAddress: string,
+  tokenID: string,
+  recipientChain: ChainId,
+  recipientAddress: Uint8Array
+): Promise<MsgExecuteContract[]> {
+  const nonce = Math.round(Math.random() * 100000);
+  return [
+    new MsgExecuteContract(
+      walletAddress,
+      tokenAddress,
+      {
+        approve: {
+          spender: tokenBridgeAddress,
+          token_id: tokenID,
+        },
+      },
+      {}
+    ),
+    new MsgExecuteContract(
+      walletAddress,
+      tokenBridgeAddress,
+      {
+        initiate_transfer: {
+          contract_addr: tokenAddress,
+          token_id: tokenID,
+          recipient_chain: recipientChain,
+          recipient: Buffer.from(recipientAddress).toString("base64"),
+          nonce: nonce,
+        },
+      },
+      {}
+    ),
+  ];
 }