+++ sdk/js/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.6
+
+Terra NFT token bridge
+getIsTransferCompleted on NFT bridge
+
 ## 0.1.5
 
 deprecated postVaaSolana